changes to implement displaying the resumes and comments from the database
