+++ cracking_the_resume/public/Javascript/viewing.js
@@ -2,17 +2,16 @@ $(document).ready(function(){
   $(".tile__img").click(function(){
     var clickedImg = $(this).attr('src');
     $("#imagePop").attr('src',clickedImg);
+    //$("#postIt").attr('name', clickedImg);
     $("#tallModal").modal("toggle");
     $("#commentModal").modal("toggle");
-
+    $("h3").text(String(clickedImg));
   })
 
 $("#closeModal").click(function(event) {
     $("#tallModal").modal("toggle");
     $("#commentModal").modal("toggle");
 });
-
-
 
   $(function () {
     $(".like").click(function () {
@@ -24,35 +23,35 @@ $("#closeModal").click(function(event) {
       input.val(input.val() - 1);
     });
 
-    var userName = "Random person";
-    var commentCount = 5;
+    // var userName = "Random person";
+    // var commentCount = 5;
 
-    function triggerButton(){
-        if(!$.trim($("#Message").val())){
-          return;
-        }
-        commentCount++;
-        $("#commentCount").text(commentCount + " comments");
+    // function triggerButton(){
+    //     if(!$.trim($("#Message").val())){
+    //       return;
+    //     }
+    //     commentCount++;
+    //     $("#commentCount").text(commentCount + " comments");
 
-        var message = $("#Message").val();
-        message = $($.parseHTML(message)).text();
-        var media = "<div class='media'>";
-        var p1 = "<p class='pull-right'><small>5 days ago</small></p>";
-        var a1 = "<a class='media-left' href='#'><img src='http://lorempixel.com/40/40/people/1/'></a>";
-        var div1 = "<div class='media-body'><h4 class='media-heading user_name'>" + userName + "</h4>" + message + "<p><small><a href=''>Like</a> - <a href=''>Share</a></small></p></div></div>";
-        $("#Message").val("");
-        $(".comments-list").append(media + p1 + a1 + div1);
+    //     var message = $("#Message").val();
+    //     message = $($.parseHTML(message)).text();
+    //     var media = "<div class='media'>";
+    //     var p1 = "<p class='pull-right'><small>5 days ago</small></p>";
+    //     var a1 = "<a class='media-left' href='#'><img src='http://lorempixel.com/40/40/people/1/'></a>";
+    //     var div1 = "<div class='media-body'><h4 class='media-heading user_name'>" + userName + "</h4>" + message + "<p><small><a href=''>Like</a> - <a href=''>Share</a></small></p></div></div>";
+    //     $("#Message").val("");
+    //     $(".comments-list").append(media + p1 + a1 + div1);
 
-    }
+    // }
 
-    $("#postIt").click(function(){
-      triggerButton();
-    });
+    // $("#postIt").click(function(){
+    //   triggerButton();
+    // });
 
-    $("#Message").keypress(function(event){
-      if(event.which == 13){
-         triggerButton();
-      }
-    });
+    // $("#Message").keypress(function(event){
+    //   if(event.which == 13){
+    //      triggerButton();
+    //   }
+    // });
   });
 });
