+++ cracking_the_resume/public/stylesheets/viewing.css
@@ -289,4 +289,9 @@ button:hover {
   left:20%;
 }
 
+.image-class{
+  height: 800px;
+  width: 600px;
+}
+
 /* sdasd */
